+++ tacotron2/data_utils.py
@@ -3,8 +3,8 @@ import numpy as np
 import torch
 import torch.utils.data
 
-import layers
-from utils import load_wav_to_torch, load_filepaths_and_text
+from tacotron2 import layers
+from tacotron2.utils import load_wav_to_torch, load_filepaths_and_text
 from text import text_to_sequence
