+++ tacotron2/logger.py
@@ -1,8 +1,9 @@
 import random
 import torch
-from tensorboardX import SummaryWriter
-from plotting_utils import plot_alignment_to_numpy, plot_spectrogram_to_numpy
-from plotting_utils import plot_gate_outputs_to_numpy
+from torch.utils.tensorboard import SummaryWriter
+
+from tacotron2.plotting_utils import plot_alignment_to_numpy, plot_spectrogram_to_numpy
+from tacotron2.plotting_utils import plot_gate_outputs_to_numpy
 
 
 class Tacotron2Logger(SummaryWriter):
