+++ tacotron2/stft.py
@@ -36,7 +36,7 @@ import torch.nn.functional as F
 from torch.autograd import Variable
 from scipy.signal import get_window
 from librosa.util import pad_center, tiny
-from audio_processing import window_sumsquare
+from tacotron2.audio_processing import window_sumsquare
 
 
 class STFT(torch.nn.Module):
